+++ packages/dynamic-flows/src/jsonSchemaForm/objectSchema/ObjectSchema.js
@@ -34,6 +34,10 @@ const ObjectSchema = (props) => {
     setModel(getValidModelParts(model, props.schema));
   }, [props.schema]);
 
+  const orderedPropertyNames = [
+    ...new Set([...(props.schema.displayOrder || []), ...Object.keys(props.schema.properties)]),
+  ];
+
   return (
     <fieldset>
       {props.schema.title && !props.hideTitle && <legend> {props.schema.title} </legend>}
@@ -42,7 +46,7 @@ const ObjectSchema = (props) => {
       {props.schema.alert && <DynamicAlert component={props.schema.alert} />}
 
       <div className="row">
-        {Object.keys(props.schema.properties).map((propertyName) => (
+        {orderedPropertyNames.map((propertyName) => (
           <div
             key={propertyName}
             className={classNames(
@@ -58,6 +62,7 @@ const ObjectSchema = (props) => {
               onChange={(newModel, schema) => onChange(propertyName, newModel, schema)}
               submitted={props.submitted}
               required={isRequired(propertyName)}
+              disabled={props.disabled}
             />
           </div>
         ))}
@@ -78,6 +83,7 @@ ObjectSchema.propTypes = {
     title: Types.string,
     description: Types.string,
     width: Types.oneOf(['sm', 'md', 'lg']),
+    displayOrder: Types.arrayOf(Types.string),
   }).isRequired,
   model: Types.shape({}),
   errors: Types.shape({}),
@@ -86,6 +92,7 @@ ObjectSchema.propTypes = {
   onChange: Types.func.isRequired,
   submitted: Types.bool.isRequired,
   hideTitle: Types.bool,
+  disabled: Types.bool,
 };
 
 ObjectSchema.defaultProps = {
@@ -94,6 +101,7 @@ ObjectSchema.defaultProps = {
   locale: 'en-GB',
   translations: {},
   hideTitle: false,
+  disabled: false,
 };
 
 export default ObjectSchema;