+++ packages/dynamic-flows/CHANGELOG.md
@@ -3,6 +3,77 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.5.0](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@5.4.1...@transferwise/dynamic-flows@5.5.0) (2020-10-07)
+
+
+### Features
+
+* Add 'disabled' prop to JsonSchemaForm [ADS-1617] ([e33fea6](https://github.com/transferwise/neptune-web/commit/e33fea6a7ee8808196c5bd4b219befa9dbaafbe7))
+
+
+
+
+
+## [5.4.1](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@5.4.0...@transferwise/dynamic-flows@5.4.1) (2020-09-30)
+
+
+### Bug Fixes
+
+* Neptune Sprint 29/30 - eslint fixes and a11y improvements ([#603](https://github.com/transferwise/neptune-web/issues/603)) ([ffcdba3](https://github.com/transferwise/neptune-web/commit/ffcdba3fff284fce8425892f15f824169f449941))
+
+
+
+
+
+# [5.4.0](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@5.3.0...@transferwise/dynamic-flows@5.4.0) (2020-09-30)
+
+
+### Features
+
+* move oneOf alerts to be below the oneOfs ([#606](https://github.com/transferwise/neptune-web/issues/606)) ([05b0433](https://github.com/transferwise/neptune-web/commit/05b04334cda333c01eb84f1cbba9cbde79bf8788))
+
+
+
+
+
+# [5.3.0](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@5.2.0...@transferwise/dynamic-flows@5.3.0) (2020-09-30)
+
+
+### Features
+
+* add logic for display order ([#601](https://github.com/transferwise/neptune-web/issues/601)) ([a42fff4](https://github.com/transferwise/neptune-web/commit/a42fff4397bb1b33de9a785314608302c4c6d6c1))
+
+
+
+
+
+# [5.2.0](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@5.1.6...@transferwise/dynamic-flows@5.2.0) (2020-09-30)
+
+
+### Features
+
+* move alert ([#605](https://github.com/transferwise/neptune-web/issues/605)) ([b3e4e30](https://github.com/transferwise/neptune-web/commit/b3e4e305fbd40265beb506d49dfa56053f7e6c1c))
+
+
+
+
+
+## [5.1.6](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@5.1.5...@transferwise/dynamic-flows@5.1.6) (2020-09-29)
+
+**Note:** Version bump only for package @transferwise/dynamic-flows
+
+
+
+
+
+## [5.1.5](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@5.1.4...@transferwise/dynamic-flows@5.1.5) (2020-09-28)
+
+**Note:** Version bump only for package @transferwise/dynamic-flows
+
+
+
+
+
 ## [5.1.4](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@5.1.3...@transferwise/dynamic-flows@5.1.4) (2020-09-24)
 
 **Note:** Version bump only for package @transferwise/dynamic-flows